Moving basic H2 cost
